agents/quantitative_auditor/pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quantitative-auditor-agent"
version = "1.0.0"
description = "Quantitative Auditor Agent - Bias Analysis"
requires-python = ">=3.10"

dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-aiplatform>=1.40.0",
    "google-genai>=0.3.0",
    "numpy>=1.24.0",
]
